Provide iiif manifests for all work types
Previously manifests were only created for GenericWorks (not for
Images).

Fixes #501
