+++ app/controllers/hyrax/generic_works_controller.rb
@@ -1,4 +1,3 @@
-require 'iiif_manifest'
 module Hyrax
   class GenericWorksController < ApplicationController
     include Hyrax::CurationConcernController
@@ -6,21 +5,7 @@ module Hyrax
     include Hyrax::WorksControllerBehavior
 
     self.curation_concern_type = GenericWork
-    self.show_presenter = Hyrax::GenericWorkShowPresenter
 
-    skip_load_and_authorize_resource only: :manifest
-
-    def manifest
-      headers['Access-Control-Allow-Origin'] = '*'
-      respond_to do |format|
-        format.json { render json: manifest_builder.to_h }
-      end
-    end
-
-    private
-
-      def manifest_builder
-        IIIFManifest::ManifestFactory.new(presenter)
-      end
+    include Hyku::IIIFManifest
   end
 end
