+++ app/controllers/hyrax/images_controller.rb
@@ -1,6 +1,3 @@
-# Generated via
-#  `rails generate curation_concerns:work Image`
-
 module Hyrax
   class ImagesController < ApplicationController
     include Hyrax::CurationConcernController
@@ -8,5 +5,7 @@ module Hyrax
     include Hyrax::WorksControllerBehavior
 
     self.curation_concern_type = Image
+
+    include Hyku::IIIFManifest
   end
 end
